requirements.txt
#devops
black==22.3.0
click==8.1.3
pytest==7.1.3
pytest-cov==4.0.0
requests==2.28.2
#rust based linter
ruff==0.0.284
#Databricks connector
databricks-sql-connector==2.1.0
python-dotenv==1.0.1

